Always check your variables!
I found and fixed some genuine problems with the toy MC.  It was generating a small number of events with pt > 1000000, which isn't going to be so great, especially when you standardize all the input values.  There were also, I think, a few cases where unlucky numbers would put you into an infinite or nearly infinite loop.  A lot of the problem seemed to stem from the fact that although the Briet-Wigner distribution mostly puts the particle masses near their usual value, there's a not-so-small chance to have the value very far away from their usual value.  So, I think once in a while, it was trying to find x values that would give a top quark with a gigantic mass, and that's not going to work.  That shouldn't happen anymore.  Furthermore, if it does go into an "infinite" loop, it should now crash after a while instead of running forever.